* dev:
  Attempt to fix stdafx.h and targetver.h encoding issues.
  Move TestGame into its own project, LittleEngine into static library. Move Assets/ to new Application root. Create convenience client-level headers. Add stdafx.h to TestGame.
  Move Utils to its own project: GameUtils. Rename namespace accordingly. Set up project dependencies (.lib and path to .h) and build order.
  Implement F12 for Debug Colliders, ` for DebugConsole
  Use variadic template Delegate<T...> and eliminate redundant Action Convert class Logger to namespace, and corresponding internal linkage globals
  Change feasible POD member functions into free functions
  Prepare EventManager; toggle Collider debug shapes.
  Add static_cast to remove warnings

+++ LittleEngine/Source/Components/SpriteAnimator.cpp
@@ -1,8 +1,8 @@
-#include "stdafx.h"
+#include "le_stdafx.h"
 #include "SpriteAnimator.h"
 #include "Levels/Level.h"
 #include "Entities/Actor.h"
-#include "Utils/Utils.h"
+#include "Utils.h"
 #include "SFMLInterface/Rendering/SpriteRenderer.h"
 #include "SFMLInterface/Rendering/RenderFactory.h"
 
@@ -61,7 +61,7 @@ namespace LittleEngine {
 		if (animating) {
 			elapsed += deltaTime;
 			Fixed ratio = Maths::Clamp01(elapsed / animTime);
-			Fixed size = sprites.size();
+			Fixed size = static_cast<int>(sprites.size());
 			index = static_cast<size_t>((ratio * size).GetInt());
 			if (index >= sprites.size()) {
 				if (looping) {
